--- backend/src/main/java/com/cbu/backend/book/bookReview/dto/response/BookReviewResponse.java
@@ -1,26 +1,0 @@
-package com.cbu.backend.bookReview.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class BookReviewResponse {
-
-    private Long id;
-
-    private Long bookId;
-    private Long writerId;
-    private String detailReview;
-    private String simpleReview;
-    private double starPoint;
-    private double starLevel;
-    private LocalDateTime createAt;
-    private LocalDateTime updateAt;
-
-
-}
