--- backend/src/main/java/com/cbu/backend/book/bookReview/repository/BookReviewRepository.java
@@ -1,7 +1,0 @@
-package com.cbu.backend.bookReview.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BookReviewRepository {
-}
