--- backend/src/main/java/com/cbu/backend/book/bookReview/service/BookReviewService.java
@@ -1,7 +1,0 @@
-package com.cbu.backend.bookReview.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BookReviewService {
-}
