+++ backend/src/main/java/com/cbu/backend/bookreview/controller/BookReviewController.java
@@ -1,7 +1,43 @@
 package com.cbu.backend.bookreview.controller;
 
-import org.springframework.web.bind.annotation.RestController;
+import com.cbu.backend.bookreview.dto.request.CreateBookReviewRequest;
+import com.cbu.backend.bookreview.service.BookReviewService;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Map;
 
 @RestController
+@RequestMapping("/bookReview")
 public class BookReviewController {
+
+    //북피뷰 등록
+    @PostMapping
+    public void enrollBookReview(@RequestBody CreateBookReviewRequest createBookReviewRequestDTO) {
+        BookReviewService bookReviewService = new BookReviewService();
+        //bookReviewService.createBookReview(createBookReviewRequestDTO);
+    }
+
+    //북리뷰 리스트 조회
+    @GetMapping
+    public void findBookReviewList() {
+    }
+
+    //북리뷰 한줄평으로 조회
+    @GetMapping
+    public void findBookReviewByTitle(@RequestParam String simpleReview) {
+    }
+
+    //북리뷰 필터 적용하여 조회
+    @GetMapping("/bookReview")
+    public void findBookReviewByFilter(@RequestParam Map<String, String> filter) {
+    }
+
+    //북리뷰 상세조회
+    @GetMapping("/bookReview/{id}")
+    public void findBookReviewDetail(@PathVariable String id) {
+    }
+
+    @DeleteMapping("/bookReview/{id}")
+    public void deleteBookReview(@PathVariable String id) {
+    }
 }
