+++ backend/src/main/java/com/cbu/backend/bookreview/dto/request/CreateBookReviewRequest.java
@@ -1,7 +1,23 @@
 package com.cbu.backend.bookreview.dto.request;
 
-import lombok.Data;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
 
-@Data
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+@AllArgsConstructor
 public class CreateBookReviewRequest {
+
+    private Long bookId;
+    private Long writerId;
+    private String detailReview;
+    private String simpleReview;
+    private double starPoint;
+    private double starLevel;
+    private LocalDateTime createAt;
+    private LocalDateTime updateAt;
+
 }
