+++ backend/src/main/java/com/cbu/backend/reply/entity/Reply.java
@@ -1,4 +1,4 @@
-package com.cbu.backend.entity;
+package com.cbu.backend.reply.entity;
 
 
 import lombok.AccessLevel;
